\name{DomainCharacterization}
\alias{DomainCharacterization}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
    Characterization of detected domains.
}
\description{
    Characterization of detected domains with three aspectes: occurence of each detected domain, cell type composition of each domain, and domain distribution of each cell type. 
}
\usage{
DomainCharacterization(domain_hclust, cell_type_distribution, k = ncol(domain_hclust[[1]]), single_cell = TRUE, sample_information_cellType = NULL, plot = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{domain_hclust}{
    A list. The output of function 'DomainHclust()'.
}
  \item{cell_type_distribution}{
    A data frame showing spatially proximate cell type distribution of each cell/spot. The row is cell/spot name. The column is cell type name. 
}
  \item{k}{
    Numeric. The number of selected domains.
}
  \item{single_cell}{
    Logical. Default is TRUE, which means the data is generated by single-cell spatial transcriptomics platform.
}
  \item{sample_information_cellType}{
    A vector showing cell type lables of each cell/spot.
}
  \item{plot}{
    Logical. Default is TRUE, which means to draw graphs.
}
}

\value{

  \item{occurence_rate }{A data frame showing occurence rate of each detected domain. This result is returned Only when 'single_cell' is TRUE.}
  \item{cell_type_average_distribution }{A data frame showing domain distribution of each cell type. This result is returned Only when 'single_cell' is TRUE.}
  \item{domain_average_distribution }{A data frame showing cell type composition of each domain.}

}
\author{
    Duan Bin
}

