\name{SpaDo-package}
\alias{SpaDo-package}
\alias{SpaDo}
\docType{package}
\title{
\packageTitle{SpaDo}
}
\description{
\packageDescription{SpaDo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpaDo}
\packageIndices{SpaDo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SpaDo}

Maintainer: \packageMaintainer{SpaDo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
