\name{SpatialReference}
\alias{SpatialReference}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Training spatial reference for domain annotation.
}
\description{
	Training spatial reference for domain annotation.
}
\usage{
	SpatialReference(cellType_distri, sample_information_region)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cellType_distri}{
	A data frame. The spatial adjacent cell type distribution of reference.
}
  \item{sample_information_region}{
	A vector showing manually annatated regions.
}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{cellType_distri}{The spatially adjacent cell type distribution of spatial reference.}
%%  \item{term_feature}{The avarage of spatially adjacent cell type distribution for each domain.}
%%  \item{cellType}{Character. Cell type names in spatially adjacent cell type distribution of spatial reference.}

}

\author{
	Bin Duan
}
